TodoAPI/pyproject.toml
[project]
name = "todoapi"
version = "0.1.0"
description = ""
authors = [
    {name = "nghiahm",email = "honghia06@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "python-environ (>=0.4.54,<0.5.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
